Emg/requirements.txt
numpy~=1.26.0
pandas~=2.2.0
matplotlib~=3.8.0
pyqtgraph~=0.13.3
scipy~=1.12.0
nidaqmx~=0.9.0
soundfile~=0.12.1
Pillow~=10.2.0
PyQt6~=6.6.1
